refactor(test): reorganize OpenAI test file structure
- Introduce a hierarchical directory structure under `test/srt/openai` to group tests by functionality:
  - `basic/`: For fundamental OpenAI API compatibility tests (completions, chat, etc.)
  - `features/`: For advanced functionality (JSON mode, hidden states, EBNF, etc.)
  - `function_call/`: For function calling and tool choice tests
  - `validation/`: For input validation and request limit tests
  - `embeddings/`: For embedding-specific server tests

- Merged `test_embedding_openai_server.py` and `TestOpenAIServerEmbedding` in `test_openai_server.py` into one
- Move all relevant OpenAI test files from `test/srt/` and the old `test/srt/openai/` into their new respective directories.

- Update all test file paths in `test/srt/run_suite.py` to match the new structure, ensuring `per-commit` and `per-commit-amd` suites are correct.

- Add previously unreferenced tests (`test_openai_server_ebnf.py`, `test_openai_server_ignore_eos.py`) to the test suites.
