+++ test/srt/openai/basic/test_serving_embedding.py
