+++ test/srt/openai/features/test_enable_thinking.py
