+++ test/srt/openai/function_call/test_tool_choice.py
