+++ test/srt/run_suite.py
@@ -29,10 +29,25 @@ suites = {
         TestFile("models/test_reward_models.py", 132),
         TestFile("models/test_vlm_models.py", 437),
         TestFile("models/test_transformers_models.py", 320),
-        TestFile("openai/test_protocol.py", 10),
-        TestFile("openai/test_serving_chat.py", 10),
-        TestFile("openai/test_serving_completions.py", 10),
-        TestFile("openai/test_serving_embedding.py", 10),
+        TestFile("openai/basic/test_protocol.py", 10),
+        TestFile("openai/basic/test_serving_chat.py", 10),
+        TestFile("openai/basic/test_serving_completions.py", 10),
+        TestFile("openai/basic/test_serving_embedding.py", 10),
+        TestFile("openai/basic/test_openai_embedding.py", 141),
+        TestFile("openai/basic/test_openai_server.py", 149),
+        TestFile("openai/features/test_cache_report.py", 100),
+        TestFile("openai/features/test_enable_thinking.py", 70),
+        TestFile("openai/features/test_json_constrained.py", 98),
+        TestFile("openai/features/test_json_mode.py", 90),
+        TestFile("openai/features/test_openai_server_ebnf.py", 95),
+        TestFile("openai/features/test_openai_server_hidden_states.py", 240),
+        TestFile("openai/features/test_reasoning_content.py", 89),
+        TestFile("openai/function_call/test_openai_function_calling.py", 60),
+        TestFile("openai/function_call/test_tool_choice.py", 226),
+        TestFile("openai/validation/test_large_max_new_tokens.py", 41),
+        TestFile("openai/validation/test_matched_stop.py", 60),
+        TestFile("openai/validation/test_openai_server_ignore_eos.py", 85),
+        TestFile("openai/validation/test_request_length_validation.py", 31),
         TestFile("test_abort.py", 51),
         TestFile("test_block_int8.py", 22),
         TestFile("test_create_kvindices.py", 2),
@@ -40,8 +55,6 @@ suites = {
         TestFile("test_eagle_infer_a.py", 370),
         TestFile("test_eagle_infer_b.py", 270),
         TestFile("test_ebnf_constrained.py", 108),
-        TestFile("test_enable_thinking.py", 70),
-        TestFile("test_embedding_openai_server.py", 141),
         TestFile("test_eval_fp8_accuracy.py", 303),
         TestFile("test_fa3.py", 376),
         # TestFile("test_flashmla.py", 352),
@@ -54,8 +67,6 @@ suites = {
         TestFile("test_int8_kernel.py", 8),
         TestFile("test_input_embeddings.py", 38),
         TestFile("test_jinja_template_utils.py", 1),
-        TestFile("test_json_constrained.py", 98),
-        TestFile("test_large_max_new_tokens.py", 41),
         TestFile("test_metrics.py", 32),
         TestFile("test_mla.py", 167),
         TestFile("test_mla_deepseek_v3.py", 342),
@@ -64,22 +75,16 @@ suites = {
         TestFile("test_mla_fp8.py", 93),
         TestFile("test_no_chunked_prefill.py", 108),
         TestFile("test_no_overlap_scheduler.py", 234),
-        TestFile("test_openai_function_calling.py", 60),
-        TestFile("test_openai_server.py", 149),
-        TestFile("test_openai_server_hidden_states.py", 240),
         TestFile("test_penalty.py", 41),
         TestFile("test_page_size.py", 60),
         TestFile("test_pytorch_sampling_backend.py", 66),
         TestFile("test_radix_attention.py", 105),
-        TestFile("test_reasoning_content.py", 89),
         TestFile("test_regex_constrained.py", 64),
-        TestFile("test_request_length_validation.py", 31),
         TestFile("test_retract_decode.py", 54),
         TestFile("test_server_args.py", 1),
         TestFile("test_skip_tokenizer_init.py", 117),
         TestFile("test_srt_engine.py", 261),
         TestFile("test_srt_endpoint.py", 130),
-        TestFile("test_tool_choice.py", 226),
         TestFile("test_torch_compile.py", 76),
         TestFile("test_torch_compile_moe.py", 172),
         TestFile("test_torch_native_attention_backend.py", 123),
@@ -107,15 +112,32 @@ suites = {
         TestFile("test_torch_compile_moe.py", 172),
         TestFile("models/test_qwen_models.py", 82),
         TestFile("models/test_reward_models.py", 132),
+        TestFile("openai/basic/test_openai_embedding.py", 141),
+        TestFile("openai/basic/test_openai_server.py", 149),
+        TestFile("openai/basic/test_protocol.py", 10),
+        TestFile("openai/basic/test_serving_chat.py", 10),
+        TestFile("openai/basic/test_serving_completions.py", 10),
+        TestFile("openai/basic/test_serving_embedding.py", 10),
         TestFile("test_abort.py", 51),
         TestFile("test_block_int8.py", 22),
         TestFile("test_create_kvindices.py", 2),
         TestFile("test_chunked_prefill.py", 313),
-        TestFile("test_embedding_openai_server.py", 141),
         TestFile("test_eval_fp8_accuracy.py", 303),
         TestFile("test_function_call_parser.py", 10),
         TestFile("test_input_embeddings.py", 38),
-        TestFile("test_large_max_new_tokens.py", 41),
+        TestFile("openai/features/test_cache_report.py", 100),
+        TestFile("openai/features/test_enable_thinking.py", 70),
+        TestFile("openai/features/test_json_constrained.py", 98),
+        TestFile("openai/features/test_json_mode.py", 90),
+        TestFile("openai/features/test_openai_server_ebnf.py", 95),
+        TestFile("openai/features/test_openai_server_hidden_states.py", 240),
+        TestFile("openai/features/test_reasoning_content.py", 89),
+        TestFile("openai/function_call/test_openai_function_calling.py", 60),
+        TestFile("openai/function_call/test_tool_choice.py", 226),
+        TestFile("openai/validation/test_large_max_new_tokens.py", 41),
+        TestFile("openai/validation/test_matched_stop.py", 60),
+        TestFile("openai/validation/test_openai_server_ignore_eos.py", 85),
+        TestFile("openai/validation/test_request_length_validation.py", 31),
         TestFile("test_metrics.py", 32),
         TestFile("test_no_chunked_prefill.py", 108),
         TestFile("test_no_overlap_scheduler.py", 234),
@@ -123,9 +145,6 @@ suites = {
         TestFile("test_page_size.py", 60),
         TestFile("test_pytorch_sampling_backend.py", 66),
         TestFile("test_radix_attention.py", 105),
-        TestFile("test_reasoning_content.py", 89),
-        TestFile("test_enable_thinking.py", 70),
-        TestFile("test_request_length_validation.py", 31),
         TestFile("test_retract_decode.py", 54),
         TestFile("test_server_args.py", 1),
         TestFile("test_skip_tokenizer_init.py", 117),
